make function to be server
