+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import Header from "@/components/header";
 import Banner from "../components/banner";
 import CookiesBanner from "@/components/cookies";
 import GoogleAnalytics from "../components/google-analytics";
-import { urlFor } from "../../util/client";
-import { getGlobalContent } from "../../util/actions";
+import { urlFor } from "../../util/actions/client";
+import { getGlobalContent } from "../../util/actions/actions";
 import "../styles/app.scss";
 import type { Metadata } from "next";
 import { cookies } from "next/headers";
