+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import "../styles/app.scss";
 import { useEffect, useState } from "react";
 import PlanningApplications from "../components/planning-application";
 import { SanityClient } from "../lib/sanityClient";
-import { Data } from "../../util/type";
+import { Data } from "../../util/helpers/type";
 import { DataClient } from "../lib/dataService";
 import ReactPaginate from "react-paginate";
 import { NextIcon } from "../../public/assets/icons";
@@ -13,8 +13,8 @@ import Input from "@/components/input";
 import { Button } from "@/components/button";
 import { ArrowIcon } from "../../public/assets/icons";
 import Link from "next/link";
-import { getLocationFromPostcode } from "../../util/geolocation";
-import { getGlobalContent } from "../../util/actions";
+import { getLocationFromPostcode } from "../../util/actions/actions";
+import { getGlobalContent } from "../../util/actions/actions";
 
 export const itemsPerPage = 6;
 const dataClient = new DataClient(new SanityClient());
