+++ src/app/planning-applications/[id]/feedback/page.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from "react";
 import Breadcrumbs from "@/components/breadcrumbs";
 import Instructions from "./instructions";
 import Questions from "../../../../components/questions";
-import { getLocalStorage } from "../../../../../util/helpLocalStorage";
-import { Data } from "../../../../../util/type";
+import { getLocalStorage } from "../../../../../util/helpers/application";
+import { Data } from "../../../../../util/helpers/type";
 
 const Feedback = () => {
   const [application, setApplication] = useState<Data>();
