+++ src/app/planning-applications/[id]/thank-you/page.tsx
@@ -3,11 +3,11 @@
 import Link from "next/link";
 import Image from "next/image";
 import { useEffect, useState } from "react";
-import { Data } from "../../../../../util/type";
-import { urlFor } from "../../../../../util/client";
-import { getLocalStorage } from "../../../../../util/helpLocalStorage";
+import { Data } from "../../../../../util/helpers/type";
+import { urlFor } from "../../../../../util/actions/client";
+import { getLocalStorage } from "../../../../../util/helpers/application";
 import Breadcrumbs from "@/components/breadcrumbs";
-import { globalContentRevalidate } from "../../../../../util/actions";
+import { globalContentRevalidate } from "../../../../../util/actions/actions";
 
 const FeedbackMessage = () => {
   const [globalConfig, setGlobalConfig] = useState<any>();
