+++ src/components/modal/index.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import { urlFor } from "../../../util/client";
+import { urlFor } from "../../../util/actions/client";
 
 const Modal = ({ setIsModalOpen, image }: any) => {
   return (
