+++ util/actions/email.ts
@@ -1,5 +1,34 @@
+"use server";
+
+import { GoogleSpreadsheet } from "google-spreadsheet";
+import { JWT } from "google-auth-library";
 import sgMail from "@sendgrid/mail";
 import { EmailJSON } from "@sendgrid/helpers/classes/email-address";
+
+export async function savefeedbackToGoogleSheet(data: any): Promise<boolean> {
+  const SPREADSHEET_ID = process.env.NEXT_PUBLIC_SPREADSHEET_ID || "";
+  const GOOGLE_CLIENT_EMAIL = process.env.NEXT_PUBLIC_GOOGLE_CLIENT_EMAIL;
+  const GOOGLE_SERVICE_PRIVATE_KEY = process.env.GOOGLE_SERVICE_PRIVATE_KEY;
+
+  try {
+    const serviceAccountAuth = new JWT({
+      email: GOOGLE_CLIENT_EMAIL,
+      key: GOOGLE_SERVICE_PRIVATE_KEY!.replace(/\\n/g, "\n"),
+      scopes: ["https://www.googleapis.com/auth/spreadsheets"],
+    });
+
+    const doc = new GoogleSpreadsheet(SPREADSHEET_ID, serviceAccountAuth);
+    await doc.loadInfo();
+
+    const sheet = doc.sheetsByIndex[0];
+    await sheet.addRow(data);
+    console.log("Google sheet saved successfully");
+    return true;
+  } catch (e) {
+    console.error("Error: ", e);
+    return false;
+  }
+}
 
 sgMail.setApiKey(process.env.SENDGRID_API_KEY!);
