+++ util/actions/validator.ts
@@ -1,6 +1,7 @@
-import { checkExistingReference } from "../util/client";
+"use server";
+import { checkExistingReference } from "./actions";
 import { z } from "zod";
-import { ValidationResult } from "../models/validationResult";
+import { ValidationResult } from "../../models/validationResult";
 
 const PlanningValidation = z.object({
   applicationNumber: z.string().min(1),
