--- util/helpLocalStorage.tsx
@@ -1,11 +1,0 @@
-export function getLocalStorage({ key, defaultValue }: any) {
-  const getValue = localStorage.getItem(key);
-
-  let storagedValue = defaultValue;
-
-  if (getValue && getValue !== "undefined") {
-    storagedValue = JSON.parse(getValue);
-  }
-
-  return storagedValue;
-}
