--- util/localStorageHelper.tsx
@@ -1,9 +1,0 @@
-export function getLocalStorage({ key, defaultValue }: any) {
-  const getValue = localStorage.getItem(key);
-  let storagedValue = defaultValue;
-  if (getValue !== null && getValue !== undefined) {
-    storagedValue = JSON.parse(getValue);
-  }
-
-  return storagedValue;
-}
