--- util/questionsInfo.tsx
@@ -1,10 +1,0 @@
-export const questions: { [key: number]: string } = {
-  3: "Design, size or height of new buildings or extensions",
-  4: "Use and function of the proposed development",
-  5: "Impacts on natural light",
-  6: "Privacy of neighbours",
-  7: "Access",
-  8: "Noise from new uses",
-  9: "Traffic, parking or road safety",
-  10: "Other",
-};
